+++ kratos/tests/test_calculate_nodal_distance_to_skin_process.py
